requirements.txt
tensorflow
numpy
matplotlib
pandas
keras
opencv-python
scikit-learn
pillow
pyyaml